커밋자 - 심우석

+++ VeriView/src/main/java/com/saoe/model/rank/RankingDTO.java
@@ -15,7 +15,7 @@ public class RankingDTO {
 	private String id;
 	private String nick;
 	private String profile;
-	private double member_score;
+	private int member_score;
 	private int grade;
 
 }
